requirements.txt
gradio==5.42.0
opencv-python>=4.8.0
requests>=2.28.0
func-timeout>=4.3.5
numpy>=1.24.0
boto3
botocore
onnxruntime
huggingface_hub>=0.16.0
Pillow>=9.0.0
